+++ Ex2/src/graph.h
@@ -11,9 +11,9 @@ void remove_vertex_by_id(int id);
 
 void remove_vertex_by_name(char* name);
 
-void add_edge_by_name(char* first_vertex_name, char* second_vertex_name, float weight);
+void add_edge_by_name(char* first_vertex_name, char* second_vertex_name, double weight);
 
-void add_edge_by_id(int head_vertex_id, int tail_vertex_id, float weight);
+void add_edge_by_id(int head_vertex_id, int tail_vertex_id, double weight);
 
 void remove_edge(int id);
